+++ ECS/ComponentBase.h
@@ -12,7 +12,7 @@ namespace Rain {
 			ComponentBase();
 			ComponentBase(int i_id);
 			virtual void Initialize() = 0;
-			virtual void Update(uint64_t i_timeSinceLastFrame) = 0;
+			virtual void Update(double i_timeSinceLastFrame) = 0;
 			virtual void Destroy() = 0;
 		};
 	}
