+++ MeshRender/MeshRenderComponent.cpp
@@ -13,7 +13,7 @@ Rain::MeshRender::MeshRenderComponent::MeshRenderComponent(int i_id, const Rende
 void Rain::MeshRender::MeshRenderComponent::Initialize() {
 
 }
-void Rain::MeshRender::MeshRenderComponent::Update(uint64_t i_timeSinceLastFrame) {
+void Rain::MeshRender::MeshRenderComponent::Update(double i_timeSinceLastFrame) {
 
 }
 void Rain::MeshRender::MeshRenderComponent::Destroy() {
